Move HTTP details to a GcsClient
