+++ crates/unftp-sbe-gcs/src/lib.rs
@@ -58,53 +58,32 @@
 // FIXME: error mapping from GCS/hyper is minimalistic, mostly PermanentError. Do proper mapping and better reporting (temporary failures too!)
 
 mod ext;
+mod gcs_client;
 pub mod object_metadata;
 pub mod options;
 mod response_body;
-mod uri;
 mod workload_identity;
 
 pub use ext::ServerExt;
 
 use async_trait::async_trait;
-use bytes::Buf;
-use futures::{prelude::*, TryStreamExt};
-use hyper::{
-    body,
-    client::connect::HttpConnector,
-    http::{header, Method, StatusCode, Uri},
-    Body, Client, Request, Response,
-};
-use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
+use gcs_client::GcsClient;
 use libunftp::{
     auth::UserDetail,
     storage::{Error, ErrorKind, Fileinfo, Metadata, StorageBackend},
 };
-use mime::APPLICATION_OCTET_STREAM;
 use object_metadata::ObjectMetadata;
 use options::AuthMethod;
-use response_body::{Item, ResponseBody};
 use std::{
     fmt::Debug,
     path::{Path, PathBuf},
-    sync::Arc,
 };
-use tokio::sync::RwLock;
-use tokio_util::codec::{BytesCodec, FramedRead};
-use uri::GcsUri;
-use yup_oauth2::ServiceAccountAuthenticator;
-
-type HttpClient = Client<HttpsConnector<HttpConnector>>;
 
 /// A [`StorageBackend`](libunftp::storage::StorageBackend) that uses Cloud storage from Google.
 /// cloned for each controlchan!
 #[derive(Clone, Debug)]
 pub struct CloudStorage {
-    uris: GcsUri,
-    client: HttpClient,
-    auth: AuthMethod,
-
-    cached_token: CachedToken,
+    gcs: GcsClient,
 }
 
 impl CloudStorage {
@@ -138,111 +117,8 @@ impl CloudStorage {
         Str: Into<String>,
         AuthHow: Into<AuthMethod>,
     {
-        let client: HttpClient = Client::builder().build(HttpsConnectorBuilder::new().with_native_roots().https_or_http().enable_http1().build());
         Self {
-            client,
-            auth: auth.into(),
-            uris: GcsUri::new(base_url.into(), bucket.into(), root),
-
-            cached_token: Default::default(),
-        }
-    }
-
-    #[tracing_attributes::instrument]
-    async fn get_token_value(&self) -> Result<String, Error> {
-        if let Some(token) = self.cached_token.get().await {
-            return Ok(token.value);
-        }
-
-        let token = self.fetch_token().await?;
-        self.cached_token.set(token.clone()).await;
-        Ok(token.value)
-    }
-
-    async fn fetch_token(&self) -> Result<Token, Error> {
-        match &self.auth {
-            AuthMethod::ServiceAccountKey(_) => {
-                let key = self.auth.to_service_account_key()?;
-                let auth = ServiceAccountAuthenticator::builder(key).hyper_client(self.client.clone()).build().await?;
-
-                auth.token(&["https://www.googleapis.com/auth/devstorage.read_write"])
-                    .map_ok(|t| t.into())
-                    .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))
-                    .await
-            }
-            AuthMethod::WorkloadIdentity(service) => workload_identity::request_token(service.clone(), self.client.clone()).await.map(|t| t.into()),
-            AuthMethod::None => Ok(Token {
-                value: "unftp_test".to_string(),
-                expires_at: None,
-            }),
-        }
-    }
-}
-
-#[derive(Default, Clone, Debug)]
-struct CachedToken {
-    inner: Arc<RwLock<Option<Token>>>,
-}
-
-impl CachedToken {
-    // get returns a token if it's available and not expired, and None otherwise.
-    async fn get(&self) -> Option<Token> {
-        let cache = self.inner.read().await;
-        cache.as_ref().and_then(|token| token.get_if_active())
-    }
-
-    async fn set(&self, token: Token) {
-        let mut cache = self.inner.write().await;
-        *cache = Some(token);
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct Token {
-    value: String,
-    expires_at: Option<time::OffsetDateTime>,
-}
-
-impl Token {
-    /// active yields true when the token is present and has not expired. In all other cases, it
-    /// returns false.
-    fn active(&self) -> bool {
-        self.expires_at
-            .map(|expires_at| {
-                let now = time::OffsetDateTime::now_utc();
-                const SAFETY_MARGIN: time::Duration = time::Duration::seconds(5);
-
-                expires_at > (now - SAFETY_MARGIN)
-            })
-            .unwrap_or(false)
-    }
-
-    fn get_if_active(&self) -> Option<Token> {
-        if self.active() {
-            Some(self.clone())
-        } else {
-            None
-        }
-    }
-}
-
-impl From<yup_oauth2::AccessToken> for Token {
-    fn from(source: yup_oauth2::AccessToken) -> Self {
-        Self {
-            value: source.as_str().to_string(),
-            expires_at: source.expiration_time(),
-        }
-    }
-}
-
-impl From<workload_identity::TokenResponse> for Token {
-    fn from(source: workload_identity::TokenResponse) -> Self {
-        let now = time::OffsetDateTime::now_utc();
-        let expires_in = time::Duration::seconds(source.expires_in.try_into().unwrap_or(0));
-
-        Self {
-            value: source.access_token,
-            expires_at: Some(now + expires_in),
+            gcs: GcsClient::new(base_url.into(), bucket.into(), root, auth),
         }
     }
 }
@@ -256,77 +132,29 @@ impl<User: UserDetail> StorageBackend<User> for CloudStorage {
     }
 
     #[tracing_attributes::instrument]
-    async fn metadata<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<Self::Metadata, Error> {
-        let uri: Uri = self.uris.metadata(path)?;
-
-        let client: HttpClient = self.client.clone();
-
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-
-        let body = unpack_response(response).await?;
-
-        let response: Item = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        response.to_metadata()
+    async fn metadata<P>(&self, _user: &User, path: P) -> Result<Self::Metadata, Error>
+    where
+        P: AsRef<Path> + Send + Debug,
+    {
+        self.gcs.item(path).await?.to_metadata()
     }
 
     async fn md5<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<String, Error>
     where
         P: AsRef<Path> + Send + Debug,
     {
-        let uri: Uri = self.uris.metadata(path)?;
-
-        let client: HttpClient = self.client.clone();
-
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-
-        let body = unpack_response(response).await?;
-
-        let response: Item = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        Ok(response.to_md5()?)
+        self.gcs.item(path).await?.to_md5()
     }
 
     #[tracing_attributes::instrument]
-    async fn list<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<Vec<Fileinfo<PathBuf, Self::Metadata>>, Error>
+    async fn list<P>(&self, _user: &User, path: P) -> Result<Vec<Fileinfo<PathBuf, Self::Metadata>>, Error>
     where
+        P: AsRef<Path> + Send + Debug,
         <Self as StorageBackend<User>>::Metadata: Metadata,
     {
-        let uri: Uri = self.uris.list(path)?;
-
-        let client: HttpClient = self.client.clone();
-
-        let token = self.get_token_value().await?;
-
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-        let body = unpack_response(response).await?;
-        let response: ResponseBody = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-        response.list()
+        self.gcs.list(path).await?.list()
     }
 
-    // #[tracing_attributes::instrument]
     async fn get_into<'a, P, W: ?Sized>(&self, user: &User, path: P, start_pos: u64, output: &'a mut W) -> Result<u64, Error>
     where
         W: tokio::io::AsyncWrite + Unpin + Sync + Send,
@@ -336,101 +164,34 @@ impl<User: UserDetail> StorageBackend<User> for CloudStorage {
         Ok(tokio::io::copy(&mut reader, output).await?)
     }
 
-    //#[tracing_attributes::instrument]
-    async fn get<P: AsRef<Path> + Send + Debug>(
-        &self,
-        _user: &User,
-        path: P,
-        start_pos: u64,
-    ) -> Result<Box<dyn tokio::io::AsyncRead + Send + Sync + Unpin>, Error> {
-        let uri: Uri = self.uris.get(path)?;
-        let client: HttpClient = self.client.clone();
-
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .header(header::RANGE, format!("bytes={}-", start_pos))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-        result_based_on_http_status(response.status(), ())?;
-
-        let futures_io_async_read = response
-            .into_body()
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))
-            .into_async_read();
-
-        let async_read = to_tokio_async_read(futures_io_async_read);
-        Ok(Box::new(async_read))
+    async fn get<P>(&self, _user: &User, path: P, start_pos: u64) -> Result<Box<dyn tokio::io::AsyncRead + Send + Sync + Unpin>, Error>
+    where
+        P: AsRef<Path> + Send + Debug,
+    {
+        self.gcs.get(path, start_pos).await
     }
 
-    async fn put<P: AsRef<Path> + Send + Debug, B: tokio::io::AsyncRead + Send + Sync + Unpin + 'static>(
-        &self,
-        _user: &User,
-        bytes: B,
-        path: P,
-        _start_pos: u64,
-    ) -> Result<u64, Error> {
-        let uri: Uri = self.uris.put(path)?;
+    async fn put<P, B>(&self, _user: &User, reader: B, path: P, _start_pos: u64) -> Result<u64, Error>
+    where
+        P: AsRef<Path> + Send + Debug,
+        B: tokio::io::AsyncRead + Send + Sync + Unpin + 'static,
+    {
+        let item = self.gcs.upload(path, reader).await?;
 
-        let client: HttpClient = self.client.clone();
-
-        let reader = tokio::io::BufReader::with_capacity(4096, bytes);
-
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .header(header::CONTENT_TYPE, APPLICATION_OCTET_STREAM.to_string())
-            .method(Method::POST)
-            .body(Body::wrap_stream(FramedRead::new(reader, BytesCodec::new()).map_ok(|b| b.freeze())))
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-        let body = unpack_response(response).await?;
-        let response: Item = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-
-        Ok(response.to_metadata()?.len())
+        Ok(item.to_metadata()?.len())
     }
 
     #[tracing_attributes::instrument]
-    async fn del<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<(), Error> {
-        let uri: Uri = self.uris.delete(path)?;
-
-        let client: HttpClient = self.client.clone();
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::DELETE)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-        unpack_response(response).await?;
-
-        Ok(())
+    async fn del<P>(&self, _user: &User, path: P) -> Result<(), Error>
+    where
+        P: AsRef<Path> + Send + Debug,
+    {
+        self.gcs.delete(path).await
     }
 
     #[tracing_attributes::instrument]
     async fn mkd<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<(), Error> {
-        let uri: Uri = self.uris.mkd(path)?;
-        let client: HttpClient = self.client.clone();
-
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .header(header::CONTENT_TYPE, APPLICATION_OCTET_STREAM.to_string())
-            .header(header::CONTENT_LENGTH, "0")
-            .method(Method::POST)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))?;
-        let response: Response<Body> = client.request(request).map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e)).await?;
-        unpack_response(response).await?;
-        Ok(())
+        self.gcs.mkd(path).await
     }
 
     #[tracing_attributes::instrument]
@@ -442,138 +203,31 @@ impl<User: UserDetail> StorageBackend<User> for CloudStorage {
     #[tracing_attributes::instrument]
     async fn rmd<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<(), Error> {
         // first call is only to figure out if the directory is actually empty or not
-        let uri: Uri = self.uris.dir_empty(&path)?;
-        let client: HttpClient = self.client.clone();
+        let path: PathBuf = path.as_ref().into();
+        let dir_empty_resp = self.gcs.dir_empty(&path).await?;
 
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))?;
-        let response: Response<Body> = client
-            .request(request)
-            .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))
-            .await?;
-        let body = unpack_response(response).await?;
-        let response: ResponseBody = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))?;
-
-        if !response.dir_exists() {
-            Err(Error::from(ErrorKind::PermanentDirectoryNotAvailable))
-        } else if !response.dir_empty() {
-            Err(Error::from(ErrorKind::PermanentDirectoryNotEmpty))
-        } else {
-            let uri: Uri = self.uris.rmd(path)?;
-            let request: Request<Body> = Request::builder()
-                .uri(uri)
-                .header(header::AUTHORIZATION, format!("Bearer {}", token))
-                .method(Method::DELETE)
-                .body(Body::empty())
-                .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))?;
-            let response: Response<Body> = client
-                .request(request)
-                .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))
-                .await?;
-            unpack_response(response).await?;
-
-            Ok(())
+        if !dir_empty_resp.dir_exists() {
+            return Err(Error::from(ErrorKind::PermanentDirectoryNotAvailable));
         }
+
+        if !dir_empty_resp.dir_empty() {
+            return Err(Error::from(ErrorKind::PermanentDirectoryNotEmpty));
+        }
+
+        self.gcs.rmd(path).await
     }
 
     #[tracing_attributes::instrument]
-    async fn cwd<P: AsRef<Path> + Send + Debug>(&self, _user: &User, path: P) -> Result<(), Error> {
-        let uri: Uri = self.uris.dir_empty(&path)?;
-        let client: HttpClient = self.client.clone();
+    async fn cwd<P>(&self, _user: &User, path: P) -> Result<(), Error>
+    where
+        P: AsRef<Path> + Send + Debug,
+    {
+        let dir_empty_resp = self.gcs.dir_empty(path).await?;
 
-        let token = self.get_token_value().await?;
-        let request: Request<Body> = Request::builder()
-            .uri(uri)
-            .header(header::AUTHORIZATION, format!("Bearer {}", token))
-            .method(Method::GET)
-            .body(Body::empty())
-            .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))?;
-        let response: Response<Body> = client
-            .request(request)
-            .map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))
-            .await?;
-        let body = unpack_response(response).await?;
-        let response: ResponseBody = serde_json::from_reader(body.reader()).map_err(|e| Error::new(ErrorKind::PermanentDirectoryNotAvailable, e))?;
-
-        if !response.dir_exists() {
+        if !dir_empty_resp.dir_exists() {
             Err(Error::from(ErrorKind::PermanentDirectoryNotAvailable))
         } else {
             Ok(())
         }
-    }
-}
-
-#[tracing_attributes::instrument]
-async fn unpack_response(response: Response<Body>) -> Result<impl Buf, Error> {
-    let status: StatusCode = response.status();
-    let body = body::aggregate(response)
-        .map_err(|e| Error::new(ErrorKind::PermanentFileNotAvailable, e))
-        .await?;
-    result_based_on_http_status(status, body)
-}
-
-fn to_tokio_async_read(r: impl futures::io::AsyncRead) -> impl tokio::io::AsyncRead {
-    tokio_util::compat::FuturesAsyncReadCompatExt::compat(r)
-}
-
-fn result_based_on_http_status<T>(status: StatusCode, ok_val: T) -> Result<T, Error> {
-    if !status.is_success() {
-        let err_kind = match status.as_u16() {
-            404 => ErrorKind::PermanentFileNotAvailable,
-            401 | 403 => ErrorKind::PermissionDenied,
-            429 => ErrorKind::TransientFileNotAvailable,
-            _ => ErrorKind::LocalError,
-        };
-        // TODO: Consume error message in body and add as error source somehow.
-        return Err(Error::from(err_kind));
-    }
-    Ok(ok_val)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn cached_token() {
-        let cache: CachedToken = Default::default();
-
-        assert_eq!(cache.get().await, None);
-
-        cache
-            .set(Token {
-                value: "the_value".to_string(),
-                expires_at: None,
-            })
-            .await;
-        assert_eq!(cache.get().await, None);
-
-        cache
-            .set(Token {
-                value: "the_value".to_string(),
-                expires_at: Some(time::OffsetDateTime::now_utc() - time::Duration::seconds(10)),
-            })
-            .await;
-        assert_eq!(cache.get().await, None);
-
-        let in_future = Some(time::OffsetDateTime::now_utc() + time::Duration::seconds(10));
-        cache
-            .set(Token {
-                value: "the_value".to_string(),
-                expires_at: in_future.clone(),
-            })
-            .await;
-        assert_eq!(
-            cache.get().await,
-            Some(Token {
-                value: "the_value".to_string(),
-                expires_at: in_future,
-            }),
-        );
     }
 }
